Unit_Testing code shifted to separate file by Mangsura
